requirements.txt
python-dotenv==0.21.0
click==8.0.3
colorama==0.4.4
Flask==2.0.3
Flask-MySQLdb==1.0.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
Werkzeug==2.0.3
